+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './body'
+export * from './cache'
 export * from './consts'
 export * from './cookie'
 export * from './request'